test: e2e-identity: move scrap_login and OidcCfg to utils/cfg.rs
We're going to remove wire_server/oidc.rs completely.
