+++ e2e-identity/tests/utils/wire_server/oidc.rs
@@ -1,14 +1,6 @@
 use http_body_util::Full;
 use hyper::body::{Bytes, Incoming};
 use hyper::{Request, Response, StatusCode};
-use scraper::Html;
-
-pub fn scrap_login(html: String) -> String {
-    let html = Html::parse_document(&html);
-    let selector = scraper::Selector::parse("form").unwrap();
-    let form = html.select(&selector).find(|_| true).unwrap();
-    form.value().attr("action").unwrap().to_string()
-}
 
 pub async fn handle_callback(req: Request<Incoming>) -> http::Result<Response<Full<Bytes>>> {
     let req_uri = req.uri().clone();
@@ -25,22 +17,4 @@ pub async fn handle_callback(req: Request<Incoming>) -> http::Result<Response<Fu
         .body(authorization_code.into())
         .unwrap();
     Ok(resp)
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct OidcCfg {
-    pub issuer: String,
-    pub authorization_endpoint: String,
-    pub token_endpoint: String,
-    pub jwks_uri: String,
-    pub userinfo_endpoint: String,
-    pub issuer_uri: Option<String>,
-}
-
-impl OidcCfg {
-    pub fn set_issuer_uri(&mut self, base: &str) {
-        let issuer_uri = url::Url::parse(&self.issuer).unwrap();
-        let issuer_uri = format!("{base}{}", issuer_uri.path());
-        self.issuer_uri = Some(issuer_uri)
-    }
 }
